+++ src/ts/checks/presets/scoreSaber.ts
@@ -79,6 +79,7 @@ const preset: InputParamsList = {
    njs: { params: { enabled: true } },
    outro: { params: { enabled: true, min: 2, max: 15 } },
    outsidePlayable: { params: { enabled: true } },
+   timingDifference: { params: { enabled: true } },
    vnjs: { params: { enabled: false } },
 };